backend/pyproject.toml
[project]
name = "xmem-backend"
version = "1.1.0"
description = "FastAPI + PostgreSQL backend for Xmem"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.115.6",
    "uvicorn[standard]",
    "sqlalchemy==2.0.36",
    "asyncpg==0.29.0",
    "psycopg2-binary",
    "alembic==1.13.2",
    "passlib[bcrypt]==1.7.4",
    "python-jose[cryptography]==3.3.0",
    "pydantic>=2.7",
    "pydantic-settings>=2.0",
    "python-multipart==0.0.17",
    "email-validator",
    "celery==5.3.4",
    "redis==5.0.1",
    "pytesseract==0.3.13",
    "pillow==11.0.0",
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "testcontainers>=4.12.0",
    "ruff>=0.9.0",
    "openai>=2.9.0",
    "httpx>=0.28.0",
    "gevent>=25.9.1",
]

[tool.uv]
package = true

[tool.setuptools]
packages = ["app"]
